chore: rename internal package from permission to rbac
This better reflects the package usage.
