+++ internal/sshportalapi/sshportal.go
@@ -8,7 +8,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/uselagoon/ssh-portal/internal/lagoondb"
-	"github.com/uselagoon/ssh-portal/internal/permission"
+	"github.com/uselagoon/ssh-portal/internal/rbac"
 	"go.opentelemetry.io/otel"
 	"go.uber.org/zap"
 )
@@ -35,7 +35,7 @@ var (
 )
 
 func sshportal(ctx context.Context, log *zap.Logger, c *nats.EncodedConn,
-	p *permission.Permission, l LagoonDBService, k KeycloakService) nats.Handler {
+	p *rbac.Permission, l LagoonDBService, k KeycloakService) nats.Handler {
 	return func(_, replySubject string, query *SSHAccessQuery) {
 		var realmRoles, userGroups []string
 		var groupProjectIDs map[string][]int
